+++ src/Domain/Entity/Construction/PartConstructionStatus.cs
