+++ src/Domain/Entity/Construction/PartInspectionStatus.cs
