+++ src/Domain/Entity/Part.cs
