+++ src/Domain/SerializableEntity/Setup/SerializableProject.cs
@@ -4,12 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Domain.Serializable_Entity.Mill
+namespace Domain.SerializableEntity.Setup
 {
     [Serializable]
-    class SerializableHeat
+    class SerializableProject
     {
-        public Guid Id { get; set; }
-        public string HeatNumber { get; set; }
+        public string MillName { get; set; }
     }
 }
